[FE][login] auth firebase connection and password validation working
